+++ AppDI321PNDIFF.lua
@@ -25,6 +25,7 @@
 
 list_acc = {}
 OUTPUT_FILE = "DI321PNDIFF"
+f_lines = nil
 
 -- false: hide internal account number -9x-
 -- true: show internal account number -9x-
@@ -39,25 +40,37 @@ function format_account(s)
     return s
 end
 
+function ReadRegistry(key, value)
+	local fi, data, content, data_type
+	
+	fi = io.popen(string.format('reg QUERY "%s" /v %s', key, value))
+	data = nil
+	if fi then
+		content = fi:read("*a")
+		data_type, data = content:match(value..'%s+(%S+)%s+(.+)\n\n')
+		fi:close()
+	end
+
+	return data_type, data
+end
+
 function format_number(v)
 	local s
-	local unary
+	local unary, sep_thousand
 	
 	if v < 0 then
-		--s = string.format("%d", math.floor(-v))
 		s = tostring(-v)
 		unary = "-"
 	else
-		--s = string.format("%d", math.floor(v))
 		s = tostring(v)
 		unary = ""
 	end
     
     local pos = string.len(s) % 3
-
+	data_type, thousand_sep = ReadRegistry('HKCU\\Control Panel\\International', 'sThousand')
+	if thousand_sep == nil then thousand_sep = '.' end
     if pos == 0 then pos = 3 end
-    return unary..string.sub(s, 1, pos).. string.gsub(string.sub(s, pos+1), "(...)", ".%1")
---	return v
+    return unary..string.sub(s, 1, pos).. string.gsub(string.sub(s, pos+1), "(...)", thousand_sep.."%1")
 end
 
 function FindFirstSeparator(line)
@@ -78,15 +91,18 @@ function FindFirstSeparator(line)
 	return sep
 end
 
-res, ReportFileName1, ReportFileName2, officer, output_sep, limit_res = iup.GetParam("Pilih Report DI321PN dalam Format CSV (Sumber: BRISIM)", nil, [=[
+res, dummy, ReportFileName1, ReportFileName2, limit_res = iup.GetParam("Pilih Report DI321 PN dalam Format CSV (Sumber: DWH)", nil, [=[
 Sumber Data: %m\n
-Report Posisi Awal: %f[OPEN|*.csv;*.txt|CURRENT|NO|NO]\n
-Report Posisi Akhir: %f[OPEN|*.csv;*.txt|CURRENT|NO|NO]\n
-PN Officer: %s{8 Digit Personal Number or leave it blank}\n
-Output Separator: %l|,|;|\n
+Report Posisi Awal: %f[OPEN|*DI321*.csv;*DI321*.gz|CURRENT|NO|NO]\n
+Report Posisi Akhir: %f[OPEN|*DI321*.csv;*DI321*.gz|CURRENT|NO|NO]\n
 Limit Result: %l|10|20|30|50|100|\n
 ]=]
-,"1. Buka Aplikasi BRISIM (https://brisim.bri.co.id)\n2. Pilih: DWH Reports\n3. Pilih: Critical Report\n4. Pilih: Table\n5. Pilih DI321(PN) - CURRENT ACCOUNT MONTHLY TRIAL BALANCE - ACTIVE(1 ROW)\n6. Download dan Save dalam format CSV", "C:\\Lua\\data\\20201231 DI321 PN PENGELOLAH.csv","C:\\Lua\\data\\20210130 DI321 PN PENGELOLAH.csv","",0,1)
+,"1. Buka Aplikasi BRISIM (https://brisim.bri.co.id)\n2. Pilih: DWH Reports\n3. Pilih: Critical Report\n4. Pilih: Table\n5. Pilih DI321(PN) - CURRENT ACCOUNT MONTHLY TRIAL BALANCE - ACTIVE(1 ROW)\n6. Download dan Save dalam format CSV", "C:\\Lua\\data\\20201231 DI321 PN PENGELOLAH.csv","C:\\Lua\\data\\20210130 DI321 PN PENGELOLAH.csv",1)
+
+data_type, output_sep = ReadRegistry('HKCU\\Control Panel\\International', 'sList')
+data_type, decimal_sep = ReadRegistry('HKCU\\Control Panel\\International', 'sDecimal')
+if output_sep == decimal_sep then output_sep =';' end
+if output_sep == nil then output_sep = ',' end
 
 if ReportFileName1 == "" or ReportFileName2 == "" then
 	print("Please select two reports to be compared")
@@ -95,15 +111,22 @@ if ReportFileName1 == "" or ReportFileName2 == "" then
 end
 
 -- convert Unicode to ANSI
+if ReportFileName1:match('%.gz$') == nil then
 print('Converting '..ReportFileName1..' to ANSI encoding')
 os.execute('type "'..ReportFileName1..'" > '..'tmp.csv')
 os.remove(ReportFileName1)
 os.rename('tmp.csv', ReportFileName1)
+f_lines = io.lines
+else
+f_lines = gzio.lines
+end
 
+if ReportFileName1:match('%.gz$') == nil then
 print('Converting '..ReportFileName2..' to ANSI encoding')
 os.execute('type "'..ReportFileName2..'" > '..'tmp.csv')
 os.remove(ReportFileName2)
 os.rename('tmp.csv', ReportFileName2)
+end
 
 -- Load first data into table list_acc
 t1 = os.clock()
@@ -127,9 +150,7 @@ for line in io.lines(ReportFileName1) do
 			acc_balance = tonumber(acc_balance)
 			acc_officer = f[17].." - "..f[18]
 			if INCLUDE_IA or (string.sub(acc_no,-3,-3) ~= "9") then
-				if (officer == "") or (f[17] == officer) then
-					list_acc[acc_no] = {acc_cif, acc_name, acc_balance, 0, -acc_balance, acc_officer}
-				end
+				list_acc[acc_no] = {acc_cif, acc_name, acc_balance, 0, -acc_balance, acc_officer}
 			end
 		end
 	end
@@ -141,7 +162,9 @@ print('Loading data from '..ReportFileName2)
 no = 1
 sep = ','
 posisi_report2 = ''
-for line in io.lines(ReportFileName2) do
+fo = io.open(OUTPUT_FILE.."_NEW.csv", "w")
+fo:write('Rekening'..output_sep..'Nama'..output_sep..'Saldo'..output_sep..'PN_Pengelola\n')
+for line in f_lines(ReportFileName2) do
 	-- only process line begin with number, skipping header
 	if no == 1 then
 		sep = FindFirstSeparator(line)
@@ -154,22 +177,26 @@ for line in io.lines(ReportFileName2) do
 			acc_name = f[8]
 			acc_balance = string.gsub(string.sub(f[10], 1, #f[10]-3), ",", "")
 			acc_balance = tonumber(acc_balance)
-			acc_officer = f[17].." - "..f[18]
+			acc_officer = f[17].."-"..f[18]
 			if INCLUDE_IA or (string.sub(acc_no,-3,-3) ~= "9") then
 				if list_acc[acc_no] then
 					list_acc[acc_no][4] = acc_balance
 					list_acc[acc_no][5] = list_acc[acc_no][4] - list_acc[acc_no][3]
 					list_acc[acc_no][6] = acc_officer
 				else
-					if (officer == "") or (f[17] == officer) then
-						list_acc[acc_no] = {acc_cif, acc_name, acc_balance, 0, -acc_balance, acc_officer}
-					end
+					list_acc[acc_no] = {acc_cif, acc_name, acc_balance, 0, -acc_balance, acc_officer}
+					fo:write(string.format('%s%s%s%s%s%s%s\n', 
+						format_account(acc_no), output_sep,
+						acc_name, output_sep,
+						format_number(acc_balance), output_sep,
+						acc_officer))
 				end
 			end
 		end
 	end
 	no = no + 1
 end
+fo:close()
 
 print('Sorting descending')
 sorted_list_acc = {}
@@ -206,7 +233,7 @@ fo2:write([=[
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
  
-<title>BRI Reporting Tool: Delta Tabungan DI321</title>
+<title>BRI Reporting Tool: Delta Giro DI321</title>
 <style type="text/css"> 
 body, html  { height: 100%; }
 html, body, div, span, applet, object, iframe,
@@ -377,6 +404,9 @@ fo2:close()
 fo:close()
 print('=== Done in '..(os.clock()-t1)..' ===')
 os.execute(OUTPUT_FILE..".htm")
---os.execute("pause")
+
+if iup.Alarm("Open List of New Created Account", "Buka file daftar rekening baru yang dibuat \ndalam periode "..posisi_report1.." sampai "..posisi_report2.." ? " ,"Ya" ,"Tidak") == 1 then
+	os.execute(OUTPUT_FILE.."_NEW.csv")
+end
